Optimize shader by splitting 3D texture into two (3D + ramp)
The 3D lookup texture was RGBA per pixel, but only the R channel was
used for the lookup. Additionally, one row of one slice of 3D texture's
G channel was used for the brighness ramp.

This both creates game data size overhead (e.g. 8x8 pattern texture is
4MB, with only 1MB being actually used data), and on the GPU
the memory bandwidth is wasted for no good reason.

Frame timings at rendering 40 layers of semitransparent quads, covering
4K resolution screen:
- NVIDIA RTX 3080Ti: 4.17 -> 4.10ms
- Intel UHD 620 (laptop GPU from 2017): 179 -> 145ms

However since now there are two textures that have to be in sync, it
is more cumbersome to setup the materials. So I've opted for storing
a "dither mode" integer property in the material, with a custom property
drawer that shows an enum (1x1, 2x2, 4x4, 8x8) and when changed sets
up both textures as needed. The actual texture properties are hidden
from the inspector.